fix conflicts for supplier

+++ BeluStore/ViewModels/SupplierViewModel.cs
@@ -1,11 +1,14 @@
 ﻿using BeluStore.Models;
 using BeluStore.Util;
+using BeluStore.Views;
 using System;
 using System.Collections.Generic;
 using System.Collections.ObjectModel;
 using System.Linq;
 using System.Text;
+using System.Text.RegularExpressions;
 using System.Threading.Tasks;
+using System.Windows;
 using System.Windows.Input;
 
 namespace BeluStore.ViewModels
@@ -41,6 +44,24 @@ namespace BeluStore.ViewModels
             }
         }
 
+
+
+        private string searchKeyword;
+        public string SearchKeyword
+        {
+            get { return searchKeyword; }
+            set
+            {
+                if (searchKeyword != value)
+                {
+                    searchKeyword = value;
+                    OnPropertyChanged();
+                }
+            }
+        }
+
+        public ICommand SearchSupplierCommand { get; set; }
+
         public ObservableCollection<Supplier> Suppliers { get; set; }
 
         public ICommand AddSupplierCommand { get; set; }
@@ -57,6 +78,7 @@ namespace BeluStore.ViewModels
             AddSupplierCommand = new RelayCommand(AddSupplier);
             DeleteSupplierCommand = new RelayCommand(DeleteSupplier);
             UpdateSupplierCommand = new RelayCommand(UpdateSupplier);
+            SearchSupplierCommand = new RelayCommand(SearchSupplier);
             ClearSupplierCommand = new RelayCommand(ClearSupplier);
         }
 
@@ -71,40 +93,53 @@ namespace BeluStore.ViewModels
 
         public void AddSupplier(object param)
         {
+
+
             if (EditedSupplier != null)
             {
-                using (var context = new BeluStoreContext())
+                bool check = ValidateInputs();
+                if (check)
                 {
-                    EditedSupplier.SupplierId = 0;
-                    context.Suppliers.Add(EditedSupplier);
-                    context.SaveChanges();
+                    using (var context = new BeluStoreContext())
+                    {
+                        EditedSupplier.SupplierId = 0;
+                        context.Suppliers.Add(EditedSupplier);
+                        context.SaveChanges();
+                        Suppliers.Add(EditedSupplier);
+                    }
 
-                    Suppliers.Add(EditedSupplier);
+                    ClearSupplier(null);
                 }
-
-                ClearSupplier(null);
             }
+
+
         }
 
         public void UpdateSupplier(object param)
         {
             if (SelectedSupplier != null && EditedSupplier != null)
             {
-                using (var context = new BeluStoreContext())
+                bool check = ValidateInputs();
+                if (check)
                 {
-                    var supplierToUpdate = context.Suppliers.Find(SelectedSupplier.SupplierId);
-                    if (supplierToUpdate != null)
+                    using (var context = new BeluStoreContext())
                     {
-                        supplierToUpdate.SupplierName = EditedSupplier.SupplierName;
-                        supplierToUpdate.SupplierPhone = EditedSupplier.SupplierPhone;
-                        supplierToUpdate.SupplierDescription = EditedSupplier.SupplierDescription;
-                        supplierToUpdate.Address = EditedSupplier.Address;
+                        var supplierToUpdate = context.Suppliers.Find(SelectedSupplier.SupplierId);
+                        if (supplierToUpdate != null)
+                        {
+                            supplierToUpdate.SupplierName = EditedSupplier.SupplierName;
+                            supplierToUpdate.SupplierPhone = EditedSupplier.SupplierPhone;
+                            supplierToUpdate.SupplierDescription = EditedSupplier.SupplierDescription;
+                            supplierToUpdate.Address = EditedSupplier.Address;
 
-                        context.SaveChanges();
+                            context.SaveChanges();
 
-                        var index = Suppliers.IndexOf(SelectedSupplier);
-                        Suppliers[index] = supplierToUpdate;
+                            var index = Suppliers.IndexOf(SelectedSupplier);
+                            Suppliers[index] = supplierToUpdate;
+                        }
                     }
+                    ClearSupplier(null);
+
                 }
             }
         }
@@ -113,27 +148,43 @@ namespace BeluStore.ViewModels
         {
             if (SelectedSupplier != null)
             {
-                using (var context = new BeluStoreContext())
-                {
-                    var supplierToDelete = context.Suppliers.Find(SelectedSupplier.SupplierId);
-                    if (supplierToDelete != null)
-                    {
-                        context.Suppliers.Remove(supplierToDelete);
-                        context.SaveChanges();
+   
+                var confirmationResult = MessageBox.Show(
+                    "Are you sure you want to delete this supplier?", 
+                    "Confirm Deletion",
+                    MessageBoxButton.YesNo, 
+                    MessageBoxImage.Warning 
+                );
 
-                        Suppliers.Remove(SelectedSupplier);
-                        ClearSupplier(null);
+                if (confirmationResult == MessageBoxResult.Yes)
+                {
+                    using (var context = new BeluStoreContext())
+                    {
+                        var supplierToDelete = context.Suppliers.Find(SelectedSupplier.SupplierId);
+                        if (supplierToDelete != null)
+                        {
+                            context.Suppliers.Remove(supplierToDelete);
+                            context.SaveChanges();
+
+                            Suppliers.Remove(SelectedSupplier);
+                            ClearSupplier(null);
+                        }
                     }
                 }
+ 
             }
         }
 
+
         public void ClearSupplier(object param)
         {
+            SearchKeyword = "";
+            SearchSupplier(SearchKeyword);
             SelectedSupplier = null;
             EditedSupplier = new Supplier();
             OnPropertyChanged(nameof(SelectedSupplier));
             OnPropertyChanged(nameof(EditedSupplier));
+
         }
 
         private Supplier CloneSupplier(Supplier supplier)
@@ -146,6 +197,53 @@ namespace BeluStore.ViewModels
                 SupplierDescription = supplier.SupplierDescription,
                 Address = supplier.Address
             };
+        }
+
+        private void SearchSupplier(object param)
+        {
+            using (var context = new BeluStoreContext())
+            {
+                var filteredSuppliers = context.Suppliers
+                    .Where(s => s.SupplierName.Contains(SearchKeyword) || s.Address.Contains(SearchKeyword) || s.SupplierPhone.Contains(SearchKeyword) || s.SupplierDescription.Contains(SearchKeyword))
+                    .ToList();
+                Suppliers = new ObservableCollection<Supplier>(filteredSuppliers);
+                OnPropertyChanged(nameof(Suppliers));
+            }
+        }
+
+        private bool ValidateInputs()
+        {
+            var errorMessages = new List<string>();
+
+            if (string.IsNullOrWhiteSpace(EditedSupplier.SupplierName) || EditedSupplier.SupplierName.Trim().Length < 5)
+            {
+                errorMessages.Add("Supplier Name must be at least 5 characters long.");
+            }
+
+            if (!Regex.IsMatch(EditedSupplier.SupplierPhone, @"^(\+?\d{10,})$"))
+            {
+                errorMessages.Add("Supplier Phone must be at least 10 digits long and contain only numbers (optionally starting with '+').");
+            }
+
+            if (string.IsNullOrWhiteSpace(EditedSupplier.SupplierDescription))
+            {
+                errorMessages.Add("Supplier Description cannot be empty.");
+            }
+
+            if (string.IsNullOrWhiteSpace(EditedSupplier.Address))
+            {
+                errorMessages.Add("Address cannot be empty.");
+            }
+
+            // Show error messages if any
+            if (errorMessages.Any())
+            {
+                MessageBox.Show(string.Join(Environment.NewLine, errorMessages), "Validation Errors", MessageBoxButton.OK, MessageBoxImage.Error);
+                return false;
+            }
+
+            // All validations passed
+            return true;
         }
     }
 }
